Uploaded June-July m6 release with compiled .jar

+++ m6/2022/june-july/src/minecraft/net/minecraft/client/gui/FontRenderer.java
@@ -360,7 +360,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, int x, int y, int color)
+    public int drawString(String text, double x, double y, int color)
     {
         return !this.enabled ? 0 : this.drawString(text, (float)x, (float)y, color, false);
     }
